+++ keras_core/layers/pooling/max_pooling_test.py
@@ -112,7 +112,7 @@ class MaxPoolingCorrectnessTest(testing.TestCase, parameterized.TestCase):
         ((2,), (2,), "valid", "channels_last"),
     )
     def test_max_pooling1d(self, pool_size, strides, padding, data_format):
-        inputs = np.arange(24, dtype=np.float).reshape((2, 3, 4))
+        inputs = np.arange(24, dtype="float32").reshape((2, 3, 4))
 
         layer = layers.MaxPooling1D(
             pool_size=pool_size,
@@ -136,7 +136,7 @@ class MaxPoolingCorrectnessTest(testing.TestCase, parameterized.TestCase):
         ((2, 3), (2, 2), "same", "channels_last"),
     )
     def test_max_pooling2d(self, pool_size, strides, padding, data_format):
-        inputs = np.arange(300, dtype=np.float).reshape((3, 5, 5, 4))
+        inputs = np.arange(300, dtype="float32").reshape((3, 5, 5, 4))
 
         layer = layers.MaxPooling2D(
             pool_size=pool_size,
@@ -161,7 +161,7 @@ class MaxPoolingCorrectnessTest(testing.TestCase, parameterized.TestCase):
         ((2, 3, 2), (2, 2, 1), "valid", "channels_last"),
     )
     def test_max_pooling3d(self, pool_size, strides, padding, data_format):
-        inputs = np.arange(240, dtype=np.float).reshape((2, 3, 4, 5, 2))
+        inputs = np.arange(240, dtype="float32").reshape((2, 3, 4, 5, 2))
 
         layer = layers.MaxPooling3D(
             pool_size=pool_size,
